combine styles into leaf
