--- src/PSADTree/Style/UserStyle.cs
@@ -1,20 +1,0 @@
-using System.DirectoryServices.AccountManagement;
-
-namespace PSADTree.Style;
-
-public sealed class UserStyle
-{
-    internal UserStyle()
-    { }
-
-    internal string GetColoredName(TreeObjectBase user)
-    {
-        return user.SamAccountName;
-    }
-
-    internal string GetColoredName(UserPrincipal user)
-    {
-        return user.SamAccountName;
-    }
-
-}
